Icons, multiple organizers and ripple effect on events
Adds three new features to events in the timeline: icons, multiple speakers and ripple effect on click. Icons are remotely controled through MEC Calendar in the wordpress page making use of a custom field, whose name is stored in , which gets joined to another string to get the appropiate ionic icon. Multiple speakers works by simply taking the speaker list and joining their names, instead of just picking the first one. Ripple effect uses ionic's builtin ripple effect component and is applied to the title and hour of the event.