Improve swiping behaviour in schedule.
Changed the swiping behaviour so the slides properly adjust to the
height of the content and swiper wrapper height so the vertical bar in
each of the dates does not look too long. Still work to do.

Also an update to CHANGELOG.md is pushed with this that adds all missing
changes to the Unreleased section