Complete notifications tab
Notifications tab now shows notifications loaded from notifications.json
in the repository root. They are displayed in a list that takes the
whole page. Clicking in a notification shows a modal that displays the
full card content as HTML and the buttons defined in the notification.
If there are no notifications yet, a whole page message with an icon and
some text is shown indicating so.