Add update checker
Adds an alert that checks for app updates when the app is opened. If
there is an update available, it will prompt the user to update, and
start a flexible update if the user decides to update and the method is
available, if not an instant updated and if none of those methods work,
opens the store.