Fix logo icons not showing in notification cards
Due to how obtaining the logo was implemented in the notification card
modals, logo icons didn't show up properly. This was caused because
these icons do not have an outline variant, but the outline was being
appended to them anyways. Now, if the icon includes the word "logo", the
"Outline" is not appended to the end of the name, thus not trying to get
the outline variant and properly showing.