Improve screen orientation lock
Up to this date, the screen orientation was locked via a Capacitor
plugin at runtime when the application was started. Now the screen
orientation is locked to portrait (allowing reversed portrait to) in the
app manifest. This will only work on Android, so for iOS another
solution will be needed.