Add ripple and fix quotation marks in timeline
Adds a ripple effect when a event is clicked in the Timeline component.
Note that it adds the ripple to the click event, not to the mousedown
one, as the second also gets triggered when sliding between dates in the
Schedule. The ripplet.js library is used for this.

Next, due to an event that had quotation marks in its title which where
not properly appearing, a map on the events is done that substitutes the
text that was appearing for quotation marks.

As a minor change, the browserlists db was updated as the
package-lock.json file reflects.