Implement event details modal
Adds a new component which renders more detailed information of an
event, and that is triggered when clicked on an event in the schedule.
The information that is shown is the content and the speakers of the
event as well as the short data (date and time, location and category),
with a link to the event in the web at the bottom.