Fix notch header padding
Adds a padding to the header to safeguard from interferences with
notches or other phone elements placed on top that interfere with the
display. This is done with the CSS env(safe-area-inset-top) variable
adding padding to the ion-toolbar element,  with a minimum padding of
10px.

Also the unused CSS selector header with the z-index is removed and the
header class on the ion-header that was not implemented.